backend-voice-app/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pydantic[email]
pyjwt
pydantic-settings
python-jose
